Add basic integration test for vino CR
This commit introduces a script to instal k8s with minukube, with
basic test environment for VINO.

And runs basic test that makes sure that if daemonset is succefully
deployed vino CR is marked as ready.

Ensure docker role is left intentionally because it also adds current
user to docker group which allows other scripts to invoke docker
commands without sudo. It will be removed in next patchsets.

Change-Id: Iff7c956aa88ffbcf0d85956da4057fce526b67fc