Add base IPAM support
ViNO needs to be able to assign IP addresses to the VMs that
it instantiates, and so needs to do some light
IP Address Management (IPAM).

This change adds a library with an in-memory implementation that
can allocate IPv4 and IPv6 addresses.  Future changes will
add persistance of IPAM state, the ability to de-allocate
subnets/ranges/ips, and additional input validation.

Change-Id: I1e2106f512f9f6fd8eb77fc032b181122158b585