Make IPAM idempotent
This makes IPAM functionality idempotent, so that is plays nicely
with the fact that the vino controller will frequently be reconciling
the input CRs that trigger IP assignment.  With this change, IPAM
will remember who asked for an IP already, and will give the same
IP back if asked again.

Change-Id: I7c6b99c2e087178a04d13bafe87909279994e26b