Add MAC Address Management
This adds MAC address management into the existing IPAM functionality
and CRD.  The ViNO CR is augmented to supply a MAC Prefix.  This prefix
is used as the first in a sequence of consecutive MAC addresses that
the IPAM code will generate when needed.

Note, there is no upper bounds checking on MAC addresses
(no End defined for the MAC Range), operating under the assumption
that the MAC addresses are for intents and purposes inexhaustable:
all RFC 1918 private MAC ranges are huge.
    x2-xx-xx-xx-xx-xx
    x6-xx-xx-xx-xx-xx
    xA-xx-xx-xx-xx-xx
    xE-xx-xx-xx-xx-xx

Change-Id: I19eb709019337acfe41acd7091ec43dc08e05648