Support proxy certs in vino
Users behind firewalls may be unable to build the vino
image as a result of OpenSSL errors stemming from certificates used
with the proxy. To address this, the images impacted have been
updated to copy the proxy certificate from vino/certs into
the image and run update-ca-certificates command. This is the same
behavior that already exists in the airshipctl and sip images.

Signed-off-by: Alexander Hughes <Alexander.Hughes@pm.me>
Change-Id: Ibef36a6967bb82868be3b5bf9860dce4a46c79cb